Add profile image
There is no way of identifying the different authors of the repository.

An identifier for each author should be used so that viewers of the
repository can have an image of each authors.

Let's add a profile picture to docs/images.

A profile picture is a great way to allow viewers to be able to
visualise each author.