Add new UI Mockup
The previous UI mockup did not have the styling we wanted our
final product to have.

Users may be confused when they see a stark difference in the UI
mockup and the actual UI.

Let's update the UI to be how we envision it to be in the completed
product.

This way, we can prevent confusion among users who download it,
expecting a different UI.