Add extra prefixes that will be used in later features
Some arguments that we would like to use in later features do not
have prefixes.

Assigning a prefix to each of these arguments allow a standardized
way of indicating these parameters in the commands.

Let's:
* Assign a prefix to each new parameter in CLI syntax.
* Modify some comments in other classes to use "BetterCallTA" instead
of "address book"

By having an idea of the parameters that we will use for future
features, we can write our code in a way that allows easy integration
of new arguments.