Fix bug in fxml files
The fxml files are not updated with the new parameters to be used.

This caused the display of Person to fail, as the fxml files did
not reflect the changes in PersonCard.

Let's:
* add studentID label
* add course label

The studentId label is added to the left of the name, separated by
a dash. It is done this way as the student ID is the unique
identifier of students, and names should not be the unique
identifier in case of duplicate names among students.