Fix incorrect response of Edit Command
The command did not work for edits that changed other properties
except ID.

This does not follow the intended purpose of the command.

Let's ensure the Edit Command can work with ID, Name, Phone, Email
and Course from the Person properties.

The other Person properties are non-essential, and will be included
in another command after the MVP version.