Change output message of Edit Command
The output message contains details for properties that were
not changed.

Aside from Id, other properties that were not changed are redundant.
These would only clutter up the output message.

Let's update the output of the Edit command to only show the
(original) Id of the person edited and the respective properties
that have been changed.

This way, the output message becomes more concise and relevant. The
Id is included to identify the person that has been edited.