Update User Guide
The format and examples used in the section on Edit command are
incorrect.

This will confuse the users as to the proper usage of the Edit command.

Let's:
* Update the format of Edit command
* Extract constraints on the format in a warning box
* Update examples

This will allow the user to quickly scan through the points they need,
whether it is the format or why their formatting may not be working.