Update ArchitectureSequenceDiagram
The input and parameter names used in the diagram are incorrect.

Users may get confused with the conflicting diagram compared to
the actual implementation of the sequence.

Let's:
* Modify the input to display the correct format for the Delete
command.
* Modify the parameter names to match the ones used for
implementation.

Reusing the diagram is optimal here since most of it is correct.
There are only a few things that need to be changed.