Change output message of Edit command
The output message uses a colon to separate both the command word
and the updated properties.

This may not be such an intuitive output message.

Let's change the separator between the command word and the rest
of the message to "-".

This way, users can differentiate between the command word and
updated fields.