Add skeleton code for Display command
Added command to display a bar chart of all students' grades in
the current filtered list.

TAs may want to have a visualisation of their students' grades.
Combining this with find allows them to narrow down which
students' grades to display.

Let's add a trivial DisplayCommand to the parser.

We should ensure that the parsing of the command works first,
before proceeding with its implementation.

PS. Updated output of ListCommand to display "students" instead
of "persons", keeping to our intended product goal.