Add UI for Histogram
There is no way to visualise the grades of students.

It may be unintuitive for TAs when they want to have a sense of
the overall progress of their students.

Let's:
* Add a class for the Histogram behaviour
* Add a placeholder for the Histogram in the UI

A low score may not always indicate that a student is falling
behind, if everyone is also falling behind. A graph will allow
TAs to visualise who is truly falling behind.

PS. Work in progress