Add dynamic display to Histogram
Histogram does not update when grades are changed or the displayed
list of students is changed.

Users will have to exit and reopen the program to see the changes
every time, which is very inconvenient.

Let's update Histogram to change with the displayed list of
students.

This way, the histogram can update as soon as changes are made,
allowing users to see the up-to-date histogram.