Fix labels in x axis of Histogram
The last label of the x axis is shown incorrectly, it shows "90-99"
instead of "90-100".

Users may be confused where grades of 100 will be categorized in
if no labels contain 100.

Let's update the last bin label to show "90-100".

This helps to clear misunderstandings with how the grade 100
is displayed in the Histogram.