Update User Guide
Some parts of the User Guide are ambiguous and confusing to users.

We should make our User Guide as clear and user-friendly as
possible, as users will refer to it to solve their problems.

Let's
* Explain some terms which may cause misunderstandings
* Standardize the format of the User Guide

This way, users can comfortably look through the User Guide for
solutions to their problem.