Fix bugs in EditCommand
The AddCommand and EditCommand do not allow names with certain symbols,
with the error message showing a problem with the email. When an
invalid person is used in Edit Command, the incorrect error message is
shown.

Incorrect error messages will confuse the user about their problem.

Let's:
* change the default email
* add a message to PersonNotFoundException

We change the default email as it takes the name and appends
"@u.nus.edu" to it, which results in an invalid email format for
names with certain symbols, such as "/".