Update Person details column of PersonCard
Long names cause the ID to be hidden by ellipses.

This can prevent users from viewing the id of that student, so they may
have to clear to remove them.

Let's set a minimum width for the Id field of the column.

This ensures the Id is always visible no matter the length of the name.